[DAC] Regenerated demo for dac_dac_wav_gen_dma_sam_e54_xpro
MH3-58446

+++ apps/dac/dac_wav_gen_dma/firmware/src/config/sam_e54_xpro/peripheral/tc/plib_tc0.c
@@ -94,7 +94,7 @@ void TC0_TimerInitialize( void )
     TC0_REGS->COUNT32.TC_INTFLAG = (uint8_t)TC_INTFLAG_Msk;
 
 
-    TC0_REGS->COUNT32.TC_EVCTRL = (uint8_t)(TC_EVCTRL_OVFEO_Msk);
+    TC0_REGS->COUNT32.TC_EVCTRL = (uint16_t)(TC_EVCTRL_OVFEO_Msk);
 
     while((TC0_REGS->COUNT32.TC_SYNCBUSY) != 0U)
     {
